requirements.txt
aubio==0.4.9
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
Flask==1.1.1
future==0.18.2
google==2.0.3
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-speech==1.3.2
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
httplib2==0.17.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
memory-profiler==0.55.0
my-voice-analysis==0.7
numpy==1.17.2
oauth2client==4.1.3
pandas==0.25.1
pocketsphinx==0.1.15
praat-parselmouth==0.3.3
protobuf==3.11.3
psutil==5.6.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pydub==0.23.1
pyloudnorm==0.1.0
PySoundFile==0.9.0.post1
python-dateutil==2.8.0
pytz==2019.2
requests==2.23.0
rsa==4.0
scipy==1.3.1
six==1.12.0
soupsieve==2.0
SpeechRecognition==3.8.1
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==0.16.1
